Motor and steering are running

+++ teensy_chibios/src/main/include/system_data.h
@@ -5,7 +5,7 @@
 
 typedef struct sensor_data_t {
    int16_t imu_angle;
-   int32_t wheel_speed;
+   int16_t wheel_speed;
    int16_t right_URF;
    int16_t left_URF;
    int16_t tof_dist;
